+++ src/Pages/Dashboard/Components/BlockSpotlight.js
@@ -93,7 +93,7 @@ const BlockSpotlight = () => {
     height,
     icon,
     moniker,
-    proposerAddress,
+    proposerAddress = '',
     time,
     validatorCount = {},
     votingPower = {},