+++ src/Pages/Stats/Stats.js
@@ -1,13 +1,10 @@
 import React from 'react';
-import Helmet from 'react-helmet';
-import { Section, Wrapper } from 'Components';
+import { Section, Header, Wrapper } from 'Components';
 import { Upgrades } from './Components';
 
 const Stats = () => (
-  <Wrapper noHeader>
-    <Helmet>
-      <title>Provenance Blockchain Explorer - Versions</title>
-    </Helmet>
+  <Wrapper>
+    <Header title="Provenance Blockchain Statistics - Versions" />
     <Section>
       <Upgrades />
     </Section>