+++ src/redux/actions/appActions.js
@@ -1,16 +1,20 @@
 import { createAction } from 'redux-actions';
-import { CHAINCODE_ID_URL } from 'consts';
+import { CHAINCODE_ID_URL, CHAINCODE_PREFIXES_URL } from 'consts';
 import { ajaxGet } from './xhrActions';
 
 // API
 export const GET_CHAINCODE_ID = 'GET_CHAINCODE_ID';
+export const GET_CHAINCODE_PREFIXES = 'GET_CHAINCODE_PREFIXES';
 // Store
 export const SET_THEME = 'SET_THEME';
 export const SET_WALLET_URL = 'SET_WALLET_URL';
 export const SET_IS_LOGGED_IN = 'SET_IS_LOGGED_IN';
 
 // API Calls
-export const getChaincodeID = () => async (dispatch) => ajaxGet(GET_CHAINCODE_ID, dispatch, CHAINCODE_ID_URL);
+export const getChaincodeID = () => async dispatch =>
+  ajaxGet(GET_CHAINCODE_ID, dispatch, CHAINCODE_ID_URL);
+export const getChaincodePrefixes = () => async dispatch =>
+  ajaxGet(GET_CHAINCODE_PREFIXES, dispatch, CHAINCODE_PREFIXES_URL);
 // Store
 export const setTheme = createAction(SET_THEME);
 export const setWalletUrl = createAction(SET_WALLET_URL);