+++ src/utils/table/formatTableData.js
@@ -305,9 +305,9 @@ export const formatTableData = (data = [], tableHeaders) => {
         // Version update events
         case 'events': {
           finalObj[dataName] = {
-            value: dataObj.skipped ? 'Skipped' : dataObj.scheduled ? 'Scheduled' : '',
-            icon: dataObj.skipped || dataObj.scheduled ? 'HELP_OUTLINE' : '',
+            value: dataObj.skipped ? '* Skipped' : dataObj.scheduled ? 'Scheduled' : '',
             hover: serverValue,
+            raw: serverValue,
           };
           break;
         }